# api/openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "x-go-type": "model.Error"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ],
        "type": "object",
        "x-go-type": "model.ErrorEnvelope"
      },
...
      "Meta": {
        "additionalProperties": false,
        "properties": {
          "claimed": {
            "type": "boolean"
          },
          "commit": {
            "type": "string"
          },
          "setup_complete": {
            "type": "boolean"
          },
          "ui_port": {
            "format": "int64",
            "type": "integer"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "claimed",
          "commit",
          "setup_complete",
          "ui_port",
          "version"
        ],
        "type": "object",
        "x-go-type": "api.Meta"
      },
...
    },
    "securitySchemes": {
      "csrfToken": {
        "description": "Double-submit echo of the non-HttpOnly `lm_csrf` cookie. Required on every non-GET of a session endpoint.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "sessionCookie": {
        "description": "`<session_id>.<secret>`; HttpOnly, SameSite=Lax, Path=/, Secure only over TLS. Only sha256(secret) is stored.",
        "in": "cookie",
        "name": "lm_session",
        "type": "apiKey"
      },
      "setupToken": {
        "description": "The one-time claim token, required from every non-loopback caller until the admin account exists (D38).",
        "in": "header",
        "name": "X-Setup-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The Llama Man management API. Generated from the route registry in internal/api; do not hand-edit (DESIGN section 3, D43).",
    "title": "Llama Man",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "operationId": "streamEvents",
        "parameters": [
          {
            "description": "Comma-separated topics to subscribe to. Empty means every topic. One of: instances, downloads, llamacpp, gpu, bench, jobs, events, notifications.",
            "in": "query",
            "name": "topics",
            "required": false,
            "schema": {
              "enum": [
                "instances",
                "downloads",
                "llamacpp",
                "gpu",
                "bench",
                "jobs",
                "events",
                "notifications"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "An event stream of `event: <topic>` / `id: <ulid>` / `data: {…}` frames, a `retry: 3000` directive and a 20 s `:keepalive` comment."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unrecognized entry in `?topics=`.",
            "x-error-codes": [
              "invalid_topic"
            ]
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No valid admin session accompanied the request.",
            "x-error-codes": [
              "unauthorized"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "This host has not been claimed yet. The SPA routes to the wizard on this code alone.",
            "x-error-codes": [
              "setup_required"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unhandled error. The message is a constant; the detail is in the journal.",
            "x-error-codes": [
              "internal_error"
            ]
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Server-Sent Events. `?topics=` filters the stream; `Last-Event-ID` replays the `events` topic from the durable log.",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/meta": {
      "get": {
        "operationId": "getMeta",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              }
            },
            "description": "This daemon's identity and claim state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unhandled error. The message is a constant; the detail is in the journal.",
            "x-error-codes": [
              "internal_error"
            ]
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The daemon is up but has not finished resolving its own identity.",
            "x-error-codes": [
              "internal_error"
            ]
          }
        },
        "security": [],
        "summary": "Version, claim state and the port the management listener landed on — what install.sh polls.",
        "tags": [
          "meta"
        ]
      }
    },
...
  },
...
}